project: Fix unparenting bugs
This adds a common disposal function that will loop through all children
and remove them on close, as long as they are descendants of He.Bin,
He.ButtonContent, and He.View. All other Model classes extend Gtk
classes which already contain disposal functions.

The only reason to add a new Disposal function is if a widget extends
Gtk.Widget itself, in which case it will have to provide it's own
disposal